WIP: convert standalone CLI tools into tokio.cli
