--- bin/__init__.py
@@ -1,1 +1,0 @@
-# This is here to provide interfaces into these scripts for unit testing
