+++ bin/summarize_tts.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-from tokio.cli.cache_collectdes import main
+from tokio.cli.summarize_tts import main
 
 if __name__ == '__main__':
     main()
