+++ tests/test_cli_darshan_bad_ost.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python
 """
-Test the bin/darshan_bad_ost.py tool
+Test the tokio.cli.darshan_bad_ost tool
 """
 
 import os
 import json
 import tokiotest
-import tokiobin.darshan_bad_ost
+import tokio.cli.darshan_bad_ost
 
 ### For tests that base all tests off of the sample Darshan log
 SAMPLE_BAD_DARSHAN_LOG = os.path.join(os.getcwd(), 'inputs', 'sample-badost.darshan')
@@ -20,26 +20,26 @@ MODEST_PVALUE_CUTOFF = "0.01"
 
 @tokiotest.needs_darshan
 def test_good_log():
-    """bin/darshan_bad_ost.py: detect no false positives in a good Darshan log
+    """cli.darshan_bad_ost: detect no false positives in a good Darshan log
     """
     tokiotest.check_darshan()
     argv = ['--json',
             "-p", MODEST_PVALUE_CUTOFF,
             SAMPLE_GOOD_DARSHAN_LOG]
-    output_str = tokiotest.run_bin(tokiobin.darshan_bad_ost, argv)
+    output_str = tokiotest.run_bin(tokio.cli.darshan_bad_ost, argv)
     decoded_result = json.loads(output_str)
     assert len(decoded_result) == 0
 
 @tokiotest.needs_darshan
 def test_bad_log():
-    """bin/darshan_bad_ost.py: detect a very bad OST
+    """cli.darshan_bad_ost: detect a very bad OST
     """
     tokiotest.check_darshan()
     argv = ['--json',
             "-p", STRONG_PVALUE_CUTOFF,
             "-c", STRONG_CORRELATION_CUTOFF,
             SAMPLE_BAD_DARSHAN_LOG]
-    output_str = tokiotest.run_bin(tokiobin.darshan_bad_ost, argv)
+    output_str = tokiotest.run_bin(tokio.cli.darshan_bad_ost, argv)
     decoded_result = json.loads(output_str)
     print("Received %d very bad OSTs:" % len(decoded_result))
     print(json.dumps(decoded_result, indent=4))
@@ -47,23 +47,23 @@ def test_bad_log():
 
 @tokiotest.needs_darshan
 def test_single_file_log():
-    """bin/darshan_bad_ost.py: handle log with insufficient data for correlation
+    """cli.darshan_bad_ost: handle log with insufficient data for correlation
     """
     tokiotest.check_darshan()
     argv = ['--json', SAMPLE_1FILE_DARSHAN_LOG]
-    output_str = tokiotest.run_bin(tokiobin.darshan_bad_ost, argv)
+    output_str = tokiotest.run_bin(tokio.cli.darshan_bad_ost, argv)
     decoded_result = json.loads(output_str)
     assert len(decoded_result) == 0
 
 @tokiotest.needs_darshan
 def test_multi_file_log():
-    """bin/darshan_bad_ost.py: correctly handle multiple input logs
+    """cli.darshan_bad_ost: correctly handle multiple input logs
     """
     tokiotest.check_darshan()
     argv = ['--json',
             '-c', MODEST_CORRELATION_CUTOFF,
             SAMPLE_BAD_DARSHAN_LOG,
             SAMPLE_GOOD_DARSHAN_LOG]
-    output_str = tokiotest.run_bin(tokiobin.darshan_bad_ost, argv)
+    output_str = tokiotest.run_bin(tokio.cli.darshan_bad_ost, argv)
     decoded_result = json.loads(output_str)
     assert len(decoded_result) == 0
