+++ tests/test_cli_darshan_scoreboard.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 """
-Test the bin/summarize_darshanlogs.py and bin/darshan_scoreboard.py tools
+Test the cli.summarize_darshanlogs and cli.darshan_scoreboard tools
 """
 
 import os
@@ -9,8 +9,8 @@ import json
 import warnings
 import nose
 import tokiotest
-import tokiobin.summarize_darshanlogs
-import tokiobin.darshan_scoreboard
+import tokio.cli.summarize_darshanlogs
+import tokio.cli.darshan_scoreboard
 
 ### For tests that base all tests off of the sample Darshan log
 SAMPLE_DARSHAN_LOGS = glob.glob(os.path.join(os.getcwd(), 'inputs', '*.darshan'))
@@ -19,78 +19,78 @@ FILTER_FOR_EXE = ['vpicio_uni', 'dbscan_read']
 
 @tokiotest.needs_darshan
 def test_input_dir():
-    """bin/summarize_darshanlogs.py with input dir
+    """cli.summarize_darshanlogs with input dir
     """
     # Need lots of error/warning suppression since our input dir contains a ton of non-Darshan logs
     warnings.filterwarnings('ignore')
     tokiotest.check_darshan()
     argv = [os.path.dirname(SAMPLE_DARSHAN_LOGS[0])]
     print("Executing: %s" % " ".join(argv))
-    output_str = tokiotest.run_bin(tokiobin.summarize_darshanlogs, argv)
+    output_str = tokiotest.run_bin(tokio.cli.summarize_darshanlogs, argv)
     decoded_result = json.loads(output_str)
     assert len(decoded_result) > 0
 
 @tokiotest.needs_darshan
 def test_input_file():
-    """bin/summarize_darshanlogs.py with one input log
+    """cli.summarize_darshanlogs with one input log
     """
     tokiotest.check_darshan()
     argv = [SAMPLE_DARSHAN_LOGS[0]]
     print("Executing: %s" % " ".join(argv))
-    output_str = tokiotest.run_bin(tokiobin.summarize_darshanlogs, argv)
+    output_str = tokiotest.run_bin(tokio.cli.summarize_darshanlogs, argv)
     decoded_result = json.loads(output_str)
     assert len(decoded_result) > 0
 
 @tokiotest.needs_darshan
 def test_input_files():
-    """bin/summarize_darshanlogs.py with multiple input logs
+    """cli.summarize_darshanlogs with multiple input logs
     """
     tokiotest.check_darshan()
     argv = SAMPLE_DARSHAN_LOGS
     print("Executing: %s" % " ".join(argv))
-    output_str = tokiotest.run_bin(tokiobin.summarize_darshanlogs, argv)
+    output_str = tokiotest.run_bin(tokio.cli.summarize_darshanlogs, argv)
     decoded_result = json.loads(output_str)
     assert len(decoded_result) > 0
 
 @tokiotest.needs_darshan
 def test_multithreaded():
-    """bin/summarize_darshanlogs.py --threads
+    """cli.summarize_darshanlogs --threads
     """
     tokiotest.check_darshan()
     argv = ['--threads', '4'] + SAMPLE_DARSHAN_LOGS
     print("Executing: %s" % " ".join(argv))
-    output_str = tokiotest.run_bin(tokiobin.summarize_darshanlogs, argv)
+    output_str = tokiotest.run_bin(tokio.cli.summarize_darshanlogs, argv)
     decoded_result = json.loads(output_str)
     assert len(decoded_result) > 0
 
 @tokiotest.needs_darshan
 @nose.tools.with_setup(tokiotest.create_tempfile, tokiotest.delete_tempfile)
 def test_scoreboard():
-    """bin/darshan_scoreboard.py ascii output
+    """cli.darshan_scoreboard ascii output
     """
     tokiotest.check_darshan()
     argv = ['--output', tokiotest.TEMP_FILE.name] + LOGS_FROM_DIR
     print("Executing: %s" % " ".join(argv))
-    tokiotest.run_bin(tokiobin.summarize_darshanlogs, argv)
+    tokiotest.run_bin(tokio.cli.summarize_darshanlogs, argv)
 
     argv = [tokiotest.TEMP_FILE.name]
     print("Executing: %s" % " ".join(argv))
-    output_str = tokiotest.run_bin(tokiobin.darshan_scoreboard, argv)
+    output_str = tokiotest.run_bin(tokio.cli.darshan_scoreboard, argv)
     assert len(output_str.splitlines()) > 5
 
 @tokiotest.needs_darshan
 @nose.tools.with_setup(tokiotest.create_tempfile, tokiotest.delete_tempfile)
 def test_scoreboard_json():
-    """bin/darshan_scoreboard.py --json
+    """cli.darshan_scoreboard --json
     """
     tokiotest.check_darshan()
     argv = ['--output', tokiotest.TEMP_FILE.name] + LOGS_FROM_DIR
     print("Executing: %s" % " ".join(argv))
-    tokiotest.run_bin(tokiobin.summarize_darshanlogs, argv)
+    tokiotest.run_bin(tokio.cli.summarize_darshanlogs, argv)
 
     argv = ['--json', tokiotest.TEMP_FILE.name]
     print("Executing: %s" % " ".join(argv))
-    output_str = tokiotest.run_bin(tokiobin.darshan_scoreboard, argv)
+    output_str = tokiotest.run_bin(tokio.cli.darshan_scoreboard, argv)
     decoded_result = json.loads(output_str)
     print("Result: %s" % decoded_result)
     assert len(decoded_result) > 0
@@ -98,16 +98,16 @@ def test_scoreboard_json():
 @tokiotest.needs_darshan
 @nose.tools.with_setup(tokiotest.create_tempfile, tokiotest.delete_tempfile)
 def test_scoreboard_limit_user():
-    """bin/darshan_scoreboard.py --limit-user
+    """cli.darshan_scoreboard --limit-user
     """
     tokiotest.check_darshan()
     argv = ['--output', tokiotest.TEMP_FILE.name] + LOGS_FROM_DIR
     print("Executing: %s" % " ".join(argv))
-    tokiotest.run_bin(tokiobin.summarize_darshanlogs, argv)
+    tokiotest.run_bin(tokio.cli.summarize_darshanlogs, argv)
 
     argv = ['--json', '--limit-user', tokiotest.SAMPLE_DARSHAN_LOG_USER, tokiotest.TEMP_FILE.name]
     print("Executing: %s" % " ".join(argv))
-    output_str = tokiotest.run_bin(tokiobin.darshan_scoreboard, argv)
+    output_str = tokiotest.run_bin(tokio.cli.darshan_scoreboard, argv)
     decoded_result = json.loads(output_str)
     print("Result: %s" % decoded_result)
     assert decoded_result['per_user']
@@ -115,16 +115,16 @@ def test_scoreboard_limit_user():
 @tokiotest.needs_darshan
 @nose.tools.with_setup(tokiotest.create_tempfile, tokiotest.delete_tempfile)
 def test_scoreboard_exclude_user():
-    """bin/darshan_scoreboard.py --exclude-user
+    """cli.darshan_scoreboard --exclude-user
     """
     tokiotest.check_darshan()
     argv = ['--output', tokiotest.TEMP_FILE.name] + LOGS_FROM_DIR
     print("Executing: %s" % " ".join(argv))
-    tokiotest.run_bin(tokiobin.summarize_darshanlogs, argv)
+    tokiotest.run_bin(tokio.cli.summarize_darshanlogs, argv)
 
     argv = ['--json', '--exclude-user', "%s" % tokiotest.SAMPLE_DARSHAN_LOG_USER, tokiotest.TEMP_FILE.name]
     print("Executing: %s" % " ".join(argv))
-    output_str = tokiotest.run_bin(tokiobin.darshan_scoreboard, argv)
+    output_str = tokiotest.run_bin(tokio.cli.darshan_scoreboard, argv)
     decoded_result = json.loads(output_str)
     print("Result: %s" % decoded_result)
     assert not decoded_result['per_user']
@@ -134,16 +134,16 @@ def test_scoreboard_exclude_user():
 @tokiotest.needs_darshan
 @nose.tools.with_setup(tokiotest.create_tempfile, tokiotest.delete_tempfile)
 def test_scoreboard_limit_fs():
-    """bin/darshan_scoreboard.py --limit-fs
+    """cli.darshan_scoreboard --limit-fs
     """
     tokiotest.check_darshan()
     argv = ['--output', tokiotest.TEMP_FILE.name] + LOGS_FROM_DIR
     print("Executing: %s" % " ".join(argv))
-    tokiotest.run_bin(tokiobin.summarize_darshanlogs, argv)
+    tokiotest.run_bin(tokio.cli.summarize_darshanlogs, argv)
 
     argv = ['--json', '--limit-fs', tokiotest.SAMPLE_DARSHAN_ALL_MOUNTS, tokiotest.TEMP_FILE.name]
     print("Executing: %s" % " ".join(argv))
-    output_str = tokiotest.run_bin(tokiobin.darshan_scoreboard, argv)
+    output_str = tokiotest.run_bin(tokio.cli.darshan_scoreboard, argv)
     decoded_result = json.loads(output_str)
     print("Result: %s" % decoded_result)
     assert decoded_result['per_user']
@@ -153,16 +153,16 @@ def test_scoreboard_limit_fs():
 @tokiotest.needs_darshan
 @nose.tools.with_setup(tokiotest.create_tempfile, tokiotest.delete_tempfile)
 def test_scoreboard_exclude_fs():
-    """bin/darshan_scoreboard.py --exclude-fs
+    """cli.darshan_scoreboard --exclude-fs
     """
     tokiotest.check_darshan()
     argv = ['--output', tokiotest.TEMP_FILE.name] + LOGS_FROM_DIR
     print("Executing: %s" % " ".join(argv))
-    tokiotest.run_bin(tokiobin.summarize_darshanlogs, argv)
+    tokiotest.run_bin(tokio.cli.summarize_darshanlogs, argv)
 
     argv = ['--json', '--exclude-fs', "%s" % tokiotest.SAMPLE_DARSHAN_ALL_MOUNTS, tokiotest.TEMP_FILE.name]
     print("Executing: %s" % " ".join(argv))
-    output_str = tokiotest.run_bin(tokiobin.darshan_scoreboard, argv)
+    output_str = tokiotest.run_bin(tokio.cli.darshan_scoreboard, argv)
     decoded_result = json.loads(output_str)
     print(decoded_result)
     assert not decoded_result['per_user']
@@ -172,16 +172,16 @@ def test_scoreboard_exclude_fs():
 @tokiotest.needs_darshan
 @nose.tools.with_setup(tokiotest.create_tempfile, tokiotest.delete_tempfile)
 def test_scoreboard_limit_fs_logical():
-    """bin/darshan_scoreboard.py --limit-fs, logical fs names
+    """cli.darshan_scoreboard --limit-fs, logical fs names
     """
     tokiotest.check_darshan()
     argv = ['--output', tokiotest.TEMP_FILE.name] + LOGS_FROM_DIR
     print("Executing: %s" % " ".join(argv))
-    tokiotest.run_bin(tokiobin.summarize_darshanlogs, argv)
+    tokiotest.run_bin(tokio.cli.summarize_darshanlogs, argv)
 
     argv = ['--json', '--limit-fs', tokiotest.SAMPLE_DARSHAN_ALL_MOUNTS_LOGICAL, tokiotest.TEMP_FILE.name]
     print("Executing: %s" % " ".join(argv))
-    output_str = tokiotest.run_bin(tokiobin.darshan_scoreboard, argv)
+    output_str = tokiotest.run_bin(tokio.cli.darshan_scoreboard, argv)
     decoded_result = json.loads(output_str)
     print("Result: %s" % decoded_result)
     assert decoded_result['per_user']
@@ -191,16 +191,16 @@ def test_scoreboard_limit_fs_logical():
 @tokiotest.needs_darshan
 @nose.tools.with_setup(tokiotest.create_tempfile, tokiotest.delete_tempfile)
 def test_scoreboard_exclude_fs_logical():
-    """bin/darshan_scoreboard.py --exclude-fs, logical fs names
+    """cli.darshan_scoreboard --exclude-fs, logical fs names
     """
     tokiotest.check_darshan()
     argv = ['--output', tokiotest.TEMP_FILE.name] + LOGS_FROM_DIR
     print("Executing: %s" % " ".join(argv))
-    tokiotest.run_bin(tokiobin.summarize_darshanlogs, argv)
+    tokiotest.run_bin(tokio.cli.summarize_darshanlogs, argv)
 
     argv = ['--json', '--exclude-fs', "%s" % tokiotest.SAMPLE_DARSHAN_ALL_MOUNTS_LOGICAL, tokiotest.TEMP_FILE.name]
     print("Executing: %s" % " ".join(argv))
-    output_str = tokiotest.run_bin(tokiobin.darshan_scoreboard, argv)
+    output_str = tokiotest.run_bin(tokio.cli.darshan_scoreboard, argv)
     decoded_result = json.loads(output_str)
     print(decoded_result)
     assert not decoded_result['per_user']
@@ -210,18 +210,18 @@ def test_scoreboard_exclude_fs_logical():
 @tokiotest.needs_darshan
 @nose.tools.with_setup(tokiotest.create_tempfile, tokiotest.delete_tempfile)
 def test_scoreboard_limit_exe():
-    """bin/darshan_scoreboard.py --limit-exe
+    """cli.darshan_scoreboard --limit-exe
     """
     tokiotest.check_darshan()
     argv = ['--output', tokiotest.TEMP_FILE.name] + LOGS_FROM_DIR
     print("Executing: %s" % " ".join(argv))
-    tokiotest.run_bin(tokiobin.summarize_darshanlogs, argv)
+    tokiotest.run_bin(tokio.cli.summarize_darshanlogs, argv)
 
     # try to get back a few specific exes that we know exist (FILTER_FOR_EXE)
     for appname in FILTER_FOR_EXE + [','.join(FILTER_FOR_EXE)]:
         argv = ['--json', '--limit-exe', appname, tokiotest.TEMP_FILE.name]
         print("Executing: %s" % " ".join(argv))
-        output_str = tokiotest.run_bin(tokiobin.darshan_scoreboard, argv)
+        output_str = tokiotest.run_bin(tokio.cli.darshan_scoreboard, argv)
         decoded_result = json.loads(output_str)
         print("Result: %s" % decoded_result)
 
@@ -236,24 +236,24 @@ def test_scoreboard_limit_exe():
 @tokiotest.needs_darshan
 @nose.tools.with_setup(tokiotest.create_tempfile, tokiotest.delete_tempfile)
 def test_scoreboard_exclude_exe():
-    """bin/darshan_scoreboard.py --exclude-exe
+    """cli.darshan_scoreboard --exclude-exe
     """
     tokiotest.check_darshan()
     argv = ['--output', tokiotest.TEMP_FILE.name] + LOGS_FROM_DIR
     print("Executing: %s" % " ".join(argv))
-    tokiotest.run_bin(tokiobin.summarize_darshanlogs, argv)
+    tokiotest.run_bin(tokio.cli.summarize_darshanlogs, argv)
 
     # get a reference dataset without anything removed
     argv = ['--json', tokiotest.TEMP_FILE.name]
     print("Executing: %s" % " ".join(argv))
-    output_str = tokiotest.run_bin(tokiobin.darshan_scoreboard, argv)
+    output_str = tokiotest.run_bin(tokio.cli.darshan_scoreboard, argv)
     reference_result = json.loads(output_str)
     print("Result: %s" % reference_result) 
 
     for appname in FILTER_FOR_EXE + [','.join(FILTER_FOR_EXE)]:
         argv = ['--json', '--exclude-exe', appname, tokiotest.TEMP_FILE.name]
         print("Executing: %s" % " ".join(argv))
-        output_str = tokiotest.run_bin(tokiobin.darshan_scoreboard, argv)
+        output_str = tokiotest.run_bin(tokio.cli.darshan_scoreboard, argv)
         decoded_result = json.loads(output_str)
         print("Result: %s" % decoded_result)
