+++ tests/test_cli_summarize_h5lmt.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 """
-Test the bin/summarize_h5lmt.py tool
+Test the cli.summarize_h5lmt tool
 """
 
 import os
@@ -8,11 +8,11 @@ import json
 import datetime
 import tokio
 import tokiotest
-import tokiobin.summarize_h5lmt
+import tokio.cli.summarize_h5lmt
 
 ### For tokio.tools.hdf5, which is used by summarize_job.py
 START_TIME = datetime.datetime.strptime(tokiotest.SAMPLE_H5LMT_DATES[0],
-                                        tokiobin.summarize_h5lmt.DATE_FMT)
+                                        tokio.cli.summarize_h5lmt.DATE_FMT)
 END_TIME = START_TIME + datetime.timedelta(days=1)
 
 INPUT_PARAMS = {
@@ -26,8 +26,8 @@ INPUT_PARAMS = {
     '--json --summary tts': ['--json', '--summary', tokiotest.SAMPLE_LMTDB_TTS_HDF5],
     'date range': [
         '--json',
-        '--start', START_TIME.strftime(tokiobin.summarize_h5lmt.DATE_FMT),
-        '--end', END_TIME.strftime(tokiobin.summarize_h5lmt.DATE_FMT),
+        '--start', START_TIME.strftime(tokio.cli.summarize_h5lmt.DATE_FMT),
+        '--end', END_TIME.strftime(tokio.cli.summarize_h5lmt.DATE_FMT),
         os.path.basename(tokiotest.SAMPLE_H5LMT_FILE)
     ],
 }
@@ -45,8 +45,8 @@ def run_summarize_h5lmt(args):
     """
     Run an instance of summarize_h5lmt.py
     """
-    print("Running %s %s" % ('bin/summarize_h5lmt.py', ' '.join(args)))
-    output_str = tokiotest.run_bin(tokiobin.summarize_h5lmt, args)
+    print("Running %s %s" % ('cli.summarize_h5lmt', ' '.join(args)))
+    output_str = tokiotest.run_bin(tokio.cli.summarize_h5lmt, args)
     assert len(output_str) > 0
 
     if '--json' in args:
@@ -61,5 +61,5 @@ def test_all():
     """
     for descr, args in INPUT_PARAMS.items():
         func = run_summarize_h5lmt
-        func.description = 'bin/summarize_h5lmt.py ' + descr
+        func.description = 'cli.summarize_h5lmt ' + descr
         yield func, args
