+++ tests/test_cli_summarize_tts.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python
 """
-Test the bin/summarize_tts_hdf5.py tool.
+Test the cli.summarize_tts tool.
 """
 
 import json
 import tokiotest
-import tokiobin.summarize_tts_hdf5
+import tokio.cli.summarize_tts
 
 INPUT_ARGS = [
     [],
@@ -24,12 +24,12 @@ def verify_json(output):
 
 def test_tts():
     """
-    bin/summarize_tts_hdf5.py
+    cli.summarize_tts
     """
     for input_type, input_file in tokiotest.SAMPLE_TIMESERIES_FILES.items():
         for args in INPUT_ARGS:
             func = exec_cmd
-            func.description = "bin/summarize_tts_hdf5.py %s (%s)" % (" ".join(args), input_type)
+            func.description = "cli.summarize_tts %s (%s)" % (" ".join(args), input_type)
             full_args = args + [input_file]
             yield func, full_args
 
@@ -38,7 +38,7 @@ def exec_cmd(argv):
     Execute the command with some arguments
     """
     print("Executing %s" % " ".join(argv))
-    output_str = tokiotest.run_bin(tokiobin.summarize_tts_hdf5, argv)
+    output_str = tokiotest.run_bin(tokio.cli.summarize_tts, argv)
 
     assert len(output_str) > 0
