+++ tokio/__init__.py
@@ -11,6 +11,7 @@ import tokio.debug
 import tokio.connectors
 import tokio.tools
 import tokio.analysis
+# import tokio.cli # do not import by default
 
 # Exceptions
 from tokio.common import ConfigError
