+++ tokio/cli/summarize_tts.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 """
 Summarize the contents of a TOKIO TimeSeries (TTS) HDF5 file generated by
 TimeSeries.commit_dataset().  This will eventually be merged with the
@@ -157,6 +156,3 @@ def main(argv=None):
             print_timestep_summary(results['timesteps'])
         if 'columns' in results:
             print_column_summary(results['columns'])
-
-if __name__ == '__main__':
-    main()
